+++ src/langchain_google_cloud_sql_mysql/loader.py
@@ -20,7 +20,7 @@ import sqlalchemy
 from langchain_community.document_loaders.base import BaseLoader
 from langchain_core.documents import Document
 
-from .mysql_engine import MySQLEngine
+from .engine import MySQLEngine
 
 DEFAULT_CONTENT_COL = "page_content"
 DEFAULT_METADATA_COL = "langchain_metadata"
